+++ src/main/scala/diplomaticobjectmodel/model/OMCaches.scala
@@ -63,35 +63,3 @@ object OMECC {
     }
   }
 }
-
-object OMCaches {
-  def dcache(p: DCacheParams, resourceBindings: Option[ResourceBindings]): OMDCache = {
-    OMDCache(
-      memoryRegions = resourceBindings.map(DiplomaticObjectModelAddressing.getOMMemoryRegions("DCache", _)).getOrElse(Nil),
-      interrupts = Nil,
-      nSets = p.nSets,
-      nWays = p.nWays,
-      blockSizeBytes = p.blockBytes,
-      dataMemorySizeBytes = p.nSets * p.nWays * p.blockBytes,
-      dataECC = p.dataECC.map(OMECC.getCode(_)),
-      tagECC = p.tagECC.map(OMECC.getCode(_)),
-      nTLBEntries = p.nTLBEntries
-    )
-  }
-
-  def icache(p: ICacheParams, resourceBindings: Option[ResourceBindings]): OMICache = {
-    OMICache(
-      memoryRegions = resourceBindings.map(DiplomaticObjectModelAddressing.getOMMemoryRegions("ICache", _)).getOrElse(Nil),
-      interrupts = Nil,
-      nSets = p.nSets,
-      nWays = p.nWays,
-      blockSizeBytes = p.blockBytes,
-      dataMemorySizeBytes = p.nSets * p.nWays * p.blockBytes,
-      dataECC = p.dataECC.map(OMECC.getCode),
-      tagECC = p.tagECC.map(OMECC.getCode),
-      nTLBEntries = p.nTLBEntries,
-      maxTimSize = p.nSets * (p.nWays-1) * p.blockBytes
-    )
-  }
-
-}
