+++ src/main/scala/subsystem/RocketSubsystem.scala
@@ -60,8 +60,8 @@ trait HasRocketTiles extends HasTiles
     val plics = plicOpt.map(_.device.getOMComponents(resourceBindingsMap))
     val clints = clintOpt.map(_.device.getOMComponents(resourceBindingsMap))
     rockets ++ plics.getOrElse(Nil) ++ clints.getOrElse(Nil)
-
   }
+
   def getOMRocketCores(resourceBindingsMap: ResourceBindingsMap): Seq[OMComponent] =
     rocketTiles.flatMap(c => c.cpuDevice.getOMComponents(resourceBindingsMap))
 }
